refactor: move PodClique FQN name helpers into the component package
Relocate GetPodCliqueNameFromPodCliqueFQN and
GetPodCliqueSetReplicaIndexFromPodCliqueFQN from internal/utils into the
component package, next to the PodClique and PodCliqueSet helpers they relate
to. This also lets the component package stop importing internal/utils.

Pure relocation. No behavior change.
